ESM Scripts (#19)
* esm scripts initial

* Add new article on ESM Scripts in PlayCanvas

This commit introduces a comprehensive blog post detailing the support for ESM (ECMAScript Modules) in PlayCanvas, highlighting its benefits, usage, and integration with existing projects. The article aims to inform developers about the modern JavaScript features available and encourages them to adopt ESM for new projects.

* Enhance ESM Scripts article with video embed and editor focus

This commit updates the blog post on ESM Scripts in PlayCanvas to emphasize the PlayCanvas editor's support for ESM Scripts. A video demonstrating the features and benefits of ESM Scripts has been added to enhance the reader's understanding. Minor text adjustments were also made for clarity.

* Refine ESM Scripts article for clarity and consistency

This commit updates the blog post on ESM Scripts in PlayCanvas, simplifying the title and section headings for better readability. Text adjustments were made to enhance clarity, and a video embed was optimized. The content now emphasizes the benefits of ESM Scripts while maintaining a focus on modern JavaScript practices.

* Update blog/2025-06-10-esm-scripts-are-here.md

Co-authored-by: Will Eastcott <will@playcanvas.com>

* Update blog/2025-06-10-esm-scripts-are-here.md

Co-authored-by: Will Eastcott <will@playcanvas.com>

* feedback

* Update blog/2025-06-10-esm-scripts-are-here.md

Co-authored-by: Will Eastcott <will@playcanvas.com>

* Refactor ESM Scripts article for improved clarity

This commit streamlines the content of the ESM Scripts blog post by removing redundant phrases and enhancing readability. The focus remains on the advantages of ESM in PlayCanvas, emphasizing cleaner code and better project scalability.

---------

Co-authored-by: Will Eastcott <will@playcanvas.com>